swicth library architecture
